+++ Sources/Sentry/include/SentryClient.h
@@ -11,7 +11,7 @@
 #import "SentryDefines.h"
 
 @class SentryOptions, SentrySession, SentryEvent, SentryScope, SentryThread, SentryEnvelope,
-    SentryFileManager;
+    SentryFileManager, SentryId;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -30,36 +30,38 @@ SENTRY_NO_INIT
 - (_Nullable instancetype)initWithOptions:(SentryOptions *)options;
 
 /**
- * Captures an SentryEvent
+ * Captures an SentryEvent.
+ * @return The SentryId of the event or SentryId.empty if the event is not sent.
  */
-- (NSString *_Nullable)captureEvent:(SentryEvent *)event
-                          withScope:(SentryScope *_Nullable)scope
-    NS_SWIFT_NAME(capture(event:scope:));
+- (SentryId *)captureEvent:(SentryEvent *)event
+                 withScope:(SentryScope *_Nullable)scope NS_SWIFT_NAME(capture(event:scope:));
 
 /**
  * Captures a NSError
+ *
+ * @return The SentryId of the event or SentryId.empty if the event is not sent.
  */
-- (NSString *_Nullable)captureError:(NSError *)error
-                          withScope:(SentryScope *_Nullable)scope
-    NS_SWIFT_NAME(capture(error:scope:));
+- (SentryId *)captureError:(NSError *)error
+                 withScope:(SentryScope *_Nullable)scope NS_SWIFT_NAME(capture(error:scope:));
 
 /**
  * Captures a NSException
+ *
+ * @return The SentryId of the event or SentryId.empty if the event is not sent.
  */
-- (NSString *_Nullable)captureException:(NSException *)exception
-                              withScope:(SentryScope *_Nullable)scope
+- (SentryId *)captureException:(NSException *)exception
+                     withScope:(SentryScope *_Nullable)scope
     NS_SWIFT_NAME(capture(exception:scope:));
 
 /**
  * Captures a Message
+ *
+ * @return The SentryId of the event or SentryId.empty if the event is not sent.
  */
-- (NSString *_Nullable)captureMessage:(NSString *)message
-                            withScope:(SentryScope *_Nullable)scope
-    NS_SWIFT_NAME(capture(message:scope:));
+- (SentryId *)captureMessage:(NSString *)message
+                   withScope:(SentryScope *_Nullable)scope NS_SWIFT_NAME(capture(message:scope:));
 
 - (void)captureSession:(SentrySession *)session NS_SWIFT_NAME(capture(session:));
-
-- (NSString *_Nullable)captureEnvelope:(SentryEnvelope *)envelope NS_SWIFT_NAME(capture(envelope:));
 
 - (SentryFileManager *)fileManager;